+++ mockredis/redis.py
@@ -1054,7 +1054,7 @@ class MockRedis(object):
         Return a suitable function from (score, member)
         """
         if withscores:
-            return lambda score_member: (score_member[1], score_cast_func(score_member[0]))
+            return lambda score_member: (score_member[1], score_cast_func(str(score_member[0])))
         else:
             return lambda score_member: score_member[1]