+++ mockredis/exceptions.py
@@ -1,5 +1,15 @@
-"Emulates exceptions raised by the Redis client"
+"""
+Emulates exceptions raised by the Redis client.
+"""
 
 
 class RedisError(Exception):
+    pass
+
+
+class ResponseError(Exception):
+    pass
+
+
+class WatchError(Exception):
     pass