+++ CHANGES.md
@@ -1,8 +1,10 @@
 Version 2.8.0.3
 
  - Support verifying unit tests against actual redis-server and redis-py.
+ - Improve exception representation/mapping.
  - Update TTL to return -2 for unknown keys.
- - Fix `zset` score_range_func behavior to expect string input
+ - Fix `zset` `score_range_func` behavior to expect string input
+ - Raise `WatchError` in `MockRedisPipeline.execute()`
 
 Version 2.8.0.2